+++ luigi/file.py
@@ -16,7 +16,7 @@ import os
 import random
 import tempfile
 import shutil
-import luigi.util
+import luigi.devutil
 from target import FileSystem, FileSystemTarget
 from luigi.format import FileWrapper
 
@@ -102,7 +102,7 @@ class File(FileSystemTarget):
         else:
             raise Exception('mode must be r/w')
 
-    @luigi.util.deprecate_kwarg('fail_if_exists', 'raise_if_exists', False)
+    @luigi.devutil.deprecate_kwarg('fail_if_exists', 'raise_if_exists', False)
     def move(self, new_path, fail_if_exists=False):
         if fail_if_exists and os.path.exists(new_path):
             raise RuntimeError('Destination exists: %s' % new_path)
@@ -117,7 +117,7 @@ class File(FileSystemTarget):
     def remove(self):
         self.fs.remove(self.path)
 
-    @luigi.util.deprecate_kwarg('fail_if_exists', 'raise_if_exists', False)
+    @luigi.devutil.deprecate_kwarg('fail_if_exists', 'raise_if_exists', False)
     def copy(self, new_path, fail_if_exists=False):
         if fail_if_exists and os.path.exists(new_path):
             raise RuntimeError('Destination exists: %s' % new_path)
