+++ luigi/mock.py
@@ -16,7 +16,7 @@ import StringIO
 import target
 import sys
 import os
-import luigi.util
+import luigi.devutil
 import multiprocessing
 
 
@@ -73,7 +73,7 @@ class MockFile(target.FileSystemTarget):
     def exists(self,):
         return self._fn in self.fs.get_all_data()
 
-    @luigi.util.deprecate_kwarg('fail_if_exists', 'raise_if_exists', False)
+    @luigi.devutil.deprecate_kwarg('fail_if_exists', 'raise_if_exists', False)
     def rename(self, path, fail_if_exists=False):
         if fail_if_exists and path in self.fs.get_all_data():
             raise RuntimeError('Destination exists: %s' % path)
